added subscription to back

+++ backend/eventmanager/src/main/java/com/polytech/eventmanager/repository/EventRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
-
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
